Move the bottom bar to a subdirectory
[#178757999]
