+++ src/geohazard-components/bottom-bar/bottom-bar-container.scss
@@ -1,4 +1,4 @@
-@import "common.scss";
+@import "../common.scss";
 
 $bottomBarBorderWidth: 1px;
 $bottomBarBorderStyle: $bottomBarBorderWidth solid #797979;
@@ -110,14 +110,14 @@ $hoverColor: #dfdfdf;
     width: 50px;
     background-size: 100%;
     cursor: pointer;
-    background-image: url("./assets/fullscreen.svg");
+    background-image: url("../assets/fullscreen.svg");
     &:hover {
-      background-image: url("./assets/fullscreen-dark.svg");
+      background-image: url("../assets/fullscreen-dark.svg");
     }
     &.fullscreen {
-      background-image: url("./assets/fullscreen-exit.svg");
+      background-image: url("../assets/fullscreen-exit.svg");
       &:hover {
-        background-image: url("./assets/fullscreen-exit-dark.svg");
+        background-image: url("../assets/fullscreen-exit-dark.svg");
       }
     }
   }
