+++ src/geohazard-components/bottom-bar/icon-button.scss
