+++ src/geohazard-components/bottom-bar/playback-controls.scss
