+++ src/geohazard-components/geohazard-mui-theme.ts
@@ -1,6 +1,6 @@
-import { createMuiTheme } from "@material-ui/core/styles";
+import { createTheme } from "@material-ui/core/styles";
 
-export default createMuiTheme({
+export default createTheme({
   palette: {
     primary: {
       main: "#aaa"
